pyproject.toml
[tool.poetry]
name = "moonikaa"
version = "1.0.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.3"
cogwatch = "^2.1.0"
python-dotenv = "^0.18.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
